refactor(DropDownMenu): Removes CSS.
Used by the widget visualisation picker. I did notice that there appears to be a prop types error in the console when opening this dropdown. It seems unrelated to this change, but not something I reproduced in my test org on SaaS. Unclear if this was introduced in this branch (perhaps as part of changes to the AutoComplete2) or another branch.